requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
kornia
scikit-learn 
